refactor(bar): cleanup
